+++ lib/element.ts
@@ -29,14 +29,16 @@ declare type TokenPredicate<T = Token> = (token: Token) => token is T;
 const getCondition = <T>(selector: string): TokenPredicate<T> => {
 	/* NOT FOR BROWSER */
 
-	if (/[^a-z\-,\s]/u.test(selector)) {
+	if (/[^a-z\-,#]/u.test(selector)) {
 		return checkToken(selector) as TokenPredicate<T>;
 	}
 
 	/* NOT FOR BROWSER END */
 
-	const types = new Set(selector.split(',').map(str => str.trim()));
-	return (({type}) => types.has(type)) as TokenPredicate<T>;
+	return (({type, name}) => selector.split(',').some(str => {
+		const [t, ...ns] = str.split('#');
+		return (!t || t === type) && ns.every(n => n === name);
+	})) as TokenPredicate<T>;
 };
 
 /** 类似HTMLElement */